+++ Android/Sunshine/S04.03-Exercise-AddMapAndSharing/app/src/main/java/com/example/android/sunshine/MainActivity.java
@@ -221,8 +221,23 @@ public class MainActivity extends AppCompatActivity implements ForecastAdapterOn
             return true;
         }
 
-        // TODO (2) Launch the map when the map menu item is clicked
+        if (id == R.id.action_map) {
+            openLocationInMap();
+            return true;
+        }
 
         return super.onOptionsItemSelected(item);
+    }
+
+    private void openLocationInMap() {
+        String addressString = "1600 Amphitheatre Parkway, Mountain View, CA";
+        Uri geoLocation = Uri.parse("geo:0,0?q=" + addressString);
+
+        Intent intent = new Intent(Intent.ACTION_VIEW, geoLocation);
+
+        if (intent.resolveActivity(getPackageManager()) != null)
+            startActivity(intent);
+        else
+            Log.d(TAG, "No apps capable of opening " + geoLocation.toString() + " installed!");
     }
 }